+++ src/API.php
@@ -9,6 +9,7 @@ use zcrmsdk\crm\crud\ZCRMRecord as Record;
 
 use zcrmsdk\crm\setup\users\ZCRMUser as User;
 use zcrmsdk\crm\exception\ZCRMException;
+use zcrmsdk\crm\setup\org\ZCRMOrganization;
 
 class API
 {
@@ -51,20 +52,24 @@ class API
             $response = $moduleInstance->getRecords($customViewId, $sortBy, $sortOrder, $page, $perPage);
 
             $records = $response->getData();
+            $info = $response->getInfo();
             $parsedRecords = self::parseRecords($records);
-
             return [
-                'records' => $parsedRecords
+                'records' => $parsedRecords,
+                'info' => [
+                    'more_records' => $info->getMoreRecords(),
+                    'count' => $info->getRecordCount(),
+                    'page' => $info->getPageNo(),
+                    'per_page' => $info->getPerPage(),
+                ],
             ];
         } catch (ZCRMException $e) {
             return [
-                'info' =>[
-                    'code' => $e->getCode(),
-                    'message' => $e->getMessage(),
-                    'exception_code' => $e->getExceptionCode(),
-                    'details' => $e->getExceptionDetails(),
-                ],
-                'records' => []
+                'code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'exception_code' => $e->getExceptionCode(),
+                'status' => 'error',
             ];
         }
     }
@@ -143,32 +148,41 @@ class API
      * @param  array  $params   [description]
      * @return [type]           [description]
      */
-    public function searchRecords($module, $word, $params = [])
+    public function searchRecords($module, $mapCriteria, $default = false, $params = [])
     {
         try {
+            $records = [];
+            $responseRecords = [];
+            $moreRecords = true;
             $page = isset($params['page']) ? $params['page'] : 1;
             $perPage = isset($params['perPage']) ? $params['perPage'] : 200;
 
-            $moduleInstance = $this->restClient->getModuleInstance($module);
-            $response = $moduleInstance->searchRecords($word, $page, $perPage);
+            $parsedCriteria = $default ? $mapCriteria : self::buildCriteria($mapCriteria);
 
-            $records = $response->getData();
-            $info = $response->getInfo();
-
-            $parsedRecords = self::parseRecords($records);
-
+            while($moreRecords) {
+                $moduleInstance = $this->restClient->getModuleInstance($module);
+                $response = $moduleInstance->searchRecordsByCriteria($parsedCriteria, $page, $perPage);
+                $records = $response->getData();
+                $requestInfo = $response->getInfo();
+                $parsedRecords = self::parseRecords($records);
+                $responseRecords = collect($responseRecords)
+                                    ->concat($parsedRecords)
+                                    ->all();
+                $moreRecords = $requestInfo->getMoreRecords();
+                $page++;
+            }
             return [
-                'records' => $parsedRecords,
+                'records' => $responseRecords,
                 'info' => [
-                    'more_records' => $info->getMoreRecords(),
-                    'count' => $info->getRecordCount(),
-                    'page' => $info->getPageNo(),
-                    'per_page' => $info->getPerPage(),
+                    'more_records' => $requestInfo->getMoreRecords(),
+                    'count' => $requestInfo->getRecordCount(),
+                    'page' => $requestInfo->getPageNo(),
+                    'per_page' => $requestInfo->getPerPage(),
                 ],
             ];
         } catch (ZCRMException $e) {
             return [
-                'info' =>[
+                'info' => [
                     'count' => 0,
                     'code' => $e->getCode(),
                     'message' => $e->getMessage(),
@@ -272,8 +286,8 @@ class API
     {
         try {
             $recordResponse = [];
-            $apiResponse = $this->restClient->getModuleInstance($module)->getRecord($id);// APIResponse Instance
-            $zcrmRecord = $apiResponse->getData();// ZCRMRecord Instance
+            $apiResponse = $this->restClient->getModuleInstance($module)->getRecord($id); // APIResponse Instance
+            $zcrmRecord = $apiResponse->getData(); // ZCRMRecord Instance
             $recordResponse = self::parseRecord($zcrmRecord);
             if ($zcrmRecord->getLineItems()) {
                 $recordResponse['Product_Details'] = self::getLineItems($zcrmRecord->getLineItems());
@@ -286,6 +300,277 @@ class API
                 'details' => $e->getExceptionDetails(),
                 'message' => $e->getMessage(),
                 'exception_code' => $e->getExceptionCode(),
+                'status' => 'error',
+            ];
+        }
+    }
+
+    /**
+     * Create Records
+     *
+     * @param String    $module         Module Name
+     * @param Array     $records        Array of records to be updated
+     * @param Array     $params         Array of parameters. Trigger [‘workflow’, ‘approval’, ‘blueprint’]
+     * @return Array    $response
+     */
+    public function createRecords($module, $records = [], $params = [])
+    {
+        try {
+            $responseRecords = [];
+            $zcrmRecords = [];
+            foreach ($records as $record) {
+                $zcrmRecord = Record::getInstance($module, null);
+                foreach ($record as $key => $value) {
+                    $zcrmRecord->setFieldValue($key, $value);
+                }
+                array_push($zcrmRecords, $zcrmRecord);
+            }
+            $moduleInstance = $this->restClient->getModuleInstance($module);
+            $bulkApiResponse = $moduleInstance->createRecords($zcrmRecords, $params);
+            $entityResponses = $bulkApiResponse->getEntityResponses();
+            foreach ($entityResponses as $entityResponse) {
+                array_push($responseRecords, $entityResponse->getResponseJSON());
+            }
+            return $responseRecords;
+        } catch (ZCRMException $e) {
+            return [
+                ‘code’ => $e->getCode(),
+                ‘details’ => $e->getExceptionDetails(),
+                ‘message’ => $e->getMessage(),
+                ‘exception_code’ => $e->getExceptionCode(),
+                ‘status’ => ‘error’,
+            ];
+        }
+    }
+
+    /**
+     * Get Attachments
+     *
+     * @param String    $module         Module Name
+     * @param Array     $id             Id of the record to fetch Attachments
+     * @param Array     $params         Additional parameters
+     *                                  Available keys: page, perPage
+     * @return Array    $response       Response in Array format
+     */
+    public function getAttachments($module, $id, $params)
+    {
+        try {
+            $page = $params['page'] ?? 1;
+            $perPage = $params['perPage'] ?? 200;
+            //
+            $recordsResponse = [];
+            $infoResponse = [];
+            //
+            $recordInstance = $this->restClient->getRecordInstance($module, $id);
+            $bulkApiResponse = $recordInstance->getAttachments($page, $perPage);
+            //
+            if ($bulkApiResponse->getData()) {
+                $zcrmAttachments = $bulkApiResponse->getData();
+                $zcrmRequestInfo = $bulkApiResponse->getInfo();
+                foreach ($zcrmAttachments as $index => $zcrmAttachment) {
+                    $recordsResponse[$index] = self::getAttachmentData($zcrmAttachment);
+                }
+                $infoResponse = [
+                    'more_records' => $zcrmRequestInfo->getMoreRecords(),
+                    'count' => $zcrmRequestInfo->getRecordCount(),
+                    'page' => $zcrmRequestInfo->getPageNo(),
+                    'per_page' => $zcrmRequestInfo->getPerPage(),
+                ];
+                return [
+                    'records' => $recordsResponse,
+                    'info' => $infoResponse,
+                ];
+            }
+            return [
+                'records' => [],
+                'info' => [
+                    'more_records' => false,
+                    'count' => 0,
+                ]
+            ];
+        } catch (ZCRMException $e) {
+            return [
+                'http_code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'code' => $e->getExceptionCode(),
+                'status' => 'error',
+            ];
+        }
+    }
+
+    /**
+     * Get Modules
+     *
+     * @return Array    $response       Response in Array format
+     */
+    public function getModules()
+    {
+
+        $modulesResponse = [];
+        $notAvailableModules = ['Visits', 'Actions_Performed'];
+
+        try {
+            $modules = $this->restClient->getAllModules()->getData();
+
+            foreach ($modules as $module) {
+                if ($module->isApiSupported() && !in_array($module->getAPIName(), $notAvailableModules)) {
+                    array_push($modulesResponse, [
+                        'id' => $module->getId(),
+                        'api_name' => $module->getAPIName(),
+                        'module_name' => $module->getModuleName(),
+                        'singular_label' => $module->getSingularLabel(),
+                        'plural_label' => $module->getPluralLabel(),
+                        'fields' => $module->getFields()
+                    ]);
+                }
+            }
+            return [
+                'records' => $modulesResponse
+            ];
+        } catch (ZCRMException $e) {
+            return [
+                'http_code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'code' => $e->getExceptionCode(),
+                'status' => 'error',
+            ];
+        }
+    }
+
+    /**
+     * Get All Profiles
+     *
+     * @param String    $orgName         Organization Name
+     * @param String     $orgId            Organization ID
+     * @return Array    $response       Response in Array format
+     */
+
+    public function getAllProfiles($orgName, $orgId)
+    {
+        try {
+            $apiResponse = [];
+            //
+            $recordInstance = $this->restClient->getOrganizationInstance($orgName, $orgId);
+            $bulkApiResponse = $recordInstance->getAllProfiles();
+            //
+            if ($bulkApiResponse->getData()) {
+                $zcrmProfiles = $bulkApiResponse->getData();
+                foreach ($zcrmProfiles as $index => $zcrmProfile) {
+                    $apiResponse[$index] = self::getProfilesData($zcrmProfile);
+                }
+
+                return [
+                    'profiles' => $apiResponse,
+                ];
+            }
+        } catch (ZCRMException $e) {
+            return [
+                'http_code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'code' => $e->getExceptionCode(),
+                'status' => 'error',
+            ];
+        }
+    }
+
+     /**
+     * Get Profile By Id
+     *
+     * @param String    $orgName         Organization Name
+     * @param String     $orgId          Organization ID
+     * @param String     $profileId      Profile ID
+     * @return Array    $response        Response in Array format
+     */
+    public function getProfileById($orgName, $orgId, $profileId)
+    {
+
+        try {
+            $recordResponse = [];
+            $apiResponse = $this->restClient->getOrganizationInstance($orgName, $orgId)->getProfile($profileId); // APIResponse Instance
+            $zcrmRecord = $apiResponse->getData(); // ZCRMRecord Instance
+            $recordResponse = self::getProfilesData($zcrmRecord);
+            return [
+                'profile_by_id' => $recordResponse
+            ];
+        } catch (ZCRMException $e) {
+            return [
+                'http_code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'code' => $e->getExceptionCode(),
+                'status' => 'error',
+            ];
+        }
+    }
+
+    /**
+     * Get Users
+     *
+     * @param String    $orgName         Organization Name
+     * @param String     $orgId          Organization ID
+     * @return Array    $response        Response in Array format
+     */
+    public function getAllUsers($orgName, $orgId)
+    {
+        try {
+            $apiResponse = [];
+            //
+            $orgIns = ZCRMOrganization::getInstance($orgName, $orgId);
+            $response = $orgIns->getAllUsers();
+            $userInstances = $response->getData();
+
+            foreach ($userInstances as $userInstance) {
+
+                if ($userInstance->getStatus() == 'active') {
+
+                    array_push(
+                        $apiResponse,
+                        self::handleUserResponse($userInstance)
+                    );
+                }
+            }
+            return $apiResponse;
+
+        } catch (ZCRMException $e) {
+            return [
+                'http_code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'code' => $e->getExceptionCode(),
+                'status' => 'error',
+            ];
+        }
+    }
+
+    /**
+     * UpdateRecord
+     *
+     * @param String    $module         Module Name
+     * @param String     $id           ID of record
+     * @return Array    $records        Response in Array format
+     */
+    public function updateRecord($module, $id, $record)
+    {
+
+        try {
+
+            $zcrmRecord = $this->restClient->getInstance()->getRecordInstance($module, $id);
+            $zcrmRecord->setFieldValue($record);
+            $apiResponse = $zcrmRecord->update();
+            $records = json_encode($apiResponse->getDetails());
+
+            return [
+                'records' => $records
+            ];
+        } catch (ZCRMException $e) {
+            return [
+                'http_code' => $e->getCode(),
+                'details' => $e->getExceptionDetails(),
+                'message' => $e->getMessage(),
+                'code' => $e->getExceptionCode(),
                 'status' => 'error',
             ];
         }