+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+## [0.6.0 - (2019-11-22)](https://github.com/InterConnecta/zoho-crm/compare/0.5.0...0.6.0)
+
+### Added
+  - [support-5.4] Add support to Laravel 5.4 ([31c356e](https://github.com/InterConnecta/zoho-crm/commit/31c356e))
+### Changed
+- [update-records] Update `updateRecords` method. ([610dd9d](https://github.com/InterConnecta/zoho-crm/commit/610dd9d),[c0c8d6e](https://github.com/InterConnecta/zoho-crm/commit/c0c8d6e),[226e43a](https://github.com/InterConnecta/zoho-crm/commit/226e43a))
+- [create-records] Update `createRecords` method. ([6344f27](https://github.com/InterConnecta/zoho-crm/commit/
+6344f27))
+- [profiles] Update methods of profiles. ([f526281](https://github.com/InterConnecta/zoho-crm/commit/f526281),[6fd9c35](https://github.com/InterConnecta/zoho-crm/commit/6fd9c35))
+- [search-records] Update `searchRecords` method. ([314b71b](https://github.com/InterConnecta/zoho-crm/commit/314b71b))
+### Removed
+- [update-records] Remove id param from `updateRecords` method ([b3e1f39](https://github.com/InterConnecta/zoho-crm/commit/b3e1f39))
+### Fixed
+- [downgrade-sdk] Set SDK version to `2.0.5`. ([924718b](https://github.com/InterConnecta/zoho-crm/commit/924718b))
+
+
+## [0.5.0 - (2019-10-14)](https://github.com/InterConnecta/zoho-crm/compare/0.4.1...0.5.0)
+
+### Added
+- [get-modules] Added getModules method ([6c86de0](https://github.com/InterConnecta/zoho-crm/commit/6c86de0))
+- getRecords ([34e9760](https://github.com/InterConnecta/zoho-crm/commit/34e9760))
+- [get-attachments] Added getAttachments method ([3634133](https://github.com/InterConnecta/zoho-crm/commit/3634133))
+- [get-attachments] Added getAttachmentData util function ([34c3e61](https://github.com/InterConnecta/zoho-crm/commit/34c3e61))
+
+
+## [0.4.1 - (2019-10-08)](https://github.com/InterConnecta/zoho-crm/compare/0.4.0...0.4.1)
+
+### Added
+  - [related-records] Get line items in Related records ([4c6d162](https://github.com/InterConnecta/zoho-crm/commit/4c6d162))
+
+
+## [0.4.0 - (2019-10-08)](https://github.com/InterConnecta/zoho-crm/compare/0.3.0...0.4.0)
+
+### Added
+  - [create-records] Added createRecords method ([5892bb4](https://github.com/InterConnecta/zoho-crm/commit/5892bb4))
+
+
 ## [0.3.0 - (2019-10-03)](https://github.com/InterConnecta/zoho-crm/compare/0.2.0...0.3.0)
 
 ### Added